SRE-Security-Automation-Scripts-main/PEM-Key-Replacement/requirements.txt
#python packages needed for this script
boto3
sys
json
time
csv
os
collections  
paramiko
botocore.exceptions